made stabable abstract for generic
