+++ Assets/Prefabs/Environment/Placeable/StabableCube.prefab
@@ -43,8 +43,8 @@ GameObject:
   - component: {fileID: 8651098137332554589}
   - component: {fileID: 855122512485786714}
   - component: {fileID: 5611420371894251611}
-  - component: {fileID: 191085177565569077}
   - component: {fileID: 3831749198362223840}
+  - component: {fileID: 694548417575640488}
   m_Layer: 10
   m_Name: StabableCube
   m_TagString: Untagged
@@ -140,22 +140,6 @@ BoxCollider:
   serializedVersion: 3
   m_Size: {x: 1, y: 1, z: 1}
   m_Center: {x: 0, y: 0, z: 0}
---- !u!114 &191085177565569077
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 1381175444881999803}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: d02e3d32b7d331243bbd140e09096fdf, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  dashStartTransform: {fileID: 4383921528496755129}
-  dashEndTransform: {fileID: 3318144330648956302}
-  dashDir: {x: 3, y: 0, z: 0}
-  dashLength: 3
 --- !u!54 &3831749198362223840
 Rigidbody:
   m_ObjectHideFlags: 0
@@ -183,6 +167,22 @@ Rigidbody:
   m_Interpolate: 0
   m_Constraints: 126
   m_CollisionDetection: 0
+--- !u!114 &694548417575640488
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1381175444881999803}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 25e23895af87ac94ba61c6b9571c1cff, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  dashStartTransform: {fileID: 4383921528496755129}
+  dashEndTransform: {fileID: 3318144330648956302}
+  dashDir: {x: 3, y: 0, z: 0}
+  dashLength: 3
 --- !u!1 &1774856901985232667
 GameObject:
   m_ObjectHideFlags: 0
