+++ Assets/Scripts/EnviornmentalFeatures/StabableEnviornment.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 25e23895af87ac94ba61c6b9571c1cff
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
